refactor: extract Menu and SideButton components
- create new API for CSS/class changes to the body/wrapper
- create context for exposing API and i18n text
- remove CSS modules since class checker requires raw class names
- temporarily breaks state persistence and reset
