+++ src/contexts/A11yContext.js
@@ -1,0 +1,14 @@
+/* eslint-disable no-unused-vars */
+import React from 'react';
+
+// default to English
+const A11yContext = React.createContext({
+  text: {},
+  bodyStyle: (prop, val) => {},
+  wrapperStyle: (prop, val) => {},
+  wrapperClass: (cname, prefix) => {},
+  restore: () => {},
+  reset: () => {},
+});
+
+export default A11yContext;
